WirelessTapeMeasure (millimeters and inches)
completed work on the wireless tape measure.  used the depthImage to
locate specific pixel in an image, then selected that pixel.
mousePressed took mouse location at that pixel and converted to 1D
array using mouseX + (mouseY*screen.width).  This number corresponds to
a higher res depth value located in the depthMap array.  that value was
extracted.  the value is actually a millimeter measurement, which
explains why the kinect has millimeter precision, because it depthMap
represents the 300,000 pixel locations but with higher res depth info
of 8000 units per pixel.  since there are supposedly 25.4 millimeters
in one inch, dividing the millimeter measurement by 25.4 converted the
original depth value to inches.  wurd