successful ClosetPointTracking!!
every draw cycle:  update(), reload the depthMap array with 300,000
integers of depth values from a call for kinect.depthMap to 1D array
[]depthValues.  then cycle through the entire array by using the
depthImage pixels for location purposes.  for each one of these rows
(going all the down to maximum y), scan the length of that row (going
all the way across to max x), and look for the pixel with the
corresponding smallest depth value.  start with the furtherest distance
as the closest depth, then swap out the closestValue with current
closestValue.  Convert each pixel location x,y coordinate into a 1D
array index location, go to that index location, read depth value,
compare that depth value to the current closest depth value.  if new
depth value is less than current, then replace current value with new
value and save the corresponding pixel location to closestX closestY.
at the end of the 300,000 pixel scan, draw the depthImage by calling
the kinect.depthImage function inside the image() function at 0,0.
then use the saved coordinates of the winner of the closest depth value
to draw an ellipse.  then repeat by returning to the top of draw and
resetting the closestValue to the furthest distance since the furthest
is the closest before proven otherwise by a pixel with a corresponding
depthMap value from its indexed location in the depthMap array.