SCOPE and the Variable Planes of Reality
Consider the variables as existing on different plans of reality within
the sketch where each plane of reality covers a different SCOPE.
Once a variable is assigned, it stays set for all the realities inside
the one in which it wsa originally assigned.
One plane of reality is global, and is therefore available everywhere.
Variables defined outside of any function are omnipresent.
Variables defined within a lower plane only exist inside that lower
plane of reality, and they DISAPPEAR whenever we leave that layer.
Like jumping from one individual consciousness to another.  Collective
consciousness exists in setup() and draw(), but individual
consciousnesses
must be REINHABITED every time we re enter the plane on which that
individual was created.  If we leave that plane, our awareness of what
is means
to be that individual consciousness is lost.  If we want to create a
variable to store info about an individual consciousness, then we have
to
declare that variable on the appropriate plane of reality.  We wouldn't
want to try to aggregate relationships between individual consciousness
with variables that build up values over multiple consciousness.  We
have different variables for that.  We declare and initialize clean, new
variables for use with each individual.  Other variables, created with
larger scope, store info about the collective consciousness by
aggregate
 info about the relationships of individuals.