lerp() AND mirrored index of the depthMap array
used variables with varying SCOPES to build up information about the
process over multiple draw cycles.
used linear interpolation function lerp();  // seems like a dynamic
scalar quantity which takes as its arguments 2 points and an increment.
Mirrored the inner most loop of a nested for loop in order to create an
array index consistent with a mirrored image.  As for loop for x counts
up to 640 by 1, create a new variable "reversedX" and assign to it the
screen width - x -1 (so that it begins at 639 and ends test false at 0.
 Then use the reversedX variable to replace the x value when
calculating the pixel index #, such that reversedX + y*640 = index # in
the depthMap array.

eliminated large abrupt  jumps by creating  a functional range to
tighten down the focus of the tracking.  BOUNDARY CONDITIONS.
Used global variables to save closestPoint as lastPoint.
used save() function